+++ internal/bridge/types.go
@@ -60,4 +60,5 @@ type ClientManager interface {
 	DisallowProxy()
 	GetAuthUpdateChannel() chan pmapi.ClientAuth
 	CheckConnection() error
+	SetUserAgent(clientName, clientVersion, os string)
 }