+++ pkg/pmapi/clientmanager.go
@@ -1,3 +1,20 @@
+// Copyright (c) 2020 Proton Technologies AG
+//
+// This file is part of ProtonMail Bridge.
+//
+// ProtonMail Bridge is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// ProtonMail Bridge is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with ProtonMail Bridge.  If not, see <https://www.gnu.org/licenses/>.
+
 package pmapi
 
 import (
@@ -96,6 +113,7 @@ func NewClientManager(config *ClientConfig) (cm *ClientManager) {
 	cm.newClient = func(userID string) Client {
 		return newClient(cm, userID)
 	}
+	cm.SetUserAgent("", "", "") // Set default user agent.
 
 	go cm.watchTokenExpirations()
 
@@ -111,6 +129,10 @@ func (cm *ClientManager) SetClientConstructor(f func(userID string) Client) {
 // SetRoundTripper sets the roundtripper used by clients created by this client manager.
 func (cm *ClientManager) SetRoundTripper(rt http.RoundTripper) {
 	cm.roundTripper = rt
+}
+
+func (cm *ClientManager) SetUserAgent(clientName, clientVersion, os string) {
+	cm.config.UserAgent = formatUserAgent(clientName, clientVersion, os)
 }
 
 // GetClient returns a client for the given userID.